+++ src/test/test_assembler.sh
@@ -47,6 +47,7 @@ AS_TEST_FILES=(
     "src/test-resources/test_shifts_16.s"
     "src/test-resources/test_new_16bit_ops.s"
     "src/test-resources/test_global_extern.s"
+    "src/test-resources/test_array.s"
 )
 
 for f in "${AS_TEST_FILES[@]}"; do